refactor: troca nome do componente e remove função inutilizada
