+++ src/components/startAlert/start-alert-style.tsx
